refactor ui base classes and resources to core ui package
