+++ app/src/main/java/com/elhady/movies/presentation/ui/mylistdetails/MyListDetailsFragment.kt
@@ -10,7 +10,7 @@ import com.elhady.movies.BR
 import com.elhady.movies.R
 import com.elhady.movies.core.common.bases.BaseFragment
 import com.elhady.movies.databinding.FragmentMyListDetailsBinding
-import com.elhady.movies.core.common.bases.SwipeToDeleteItem
+import com.elhady.movies.core.ui.bases.SwipeToDeleteItem
 import com.elhady.movies.presentation.viewmodel.mylistdetails.MyListDetailsUiEvent
 import com.elhady.movies.presentation.viewmodel.mylistdetails.MyListDetailsUiState
 import com.elhady.movies.presentation.viewmodel.mylistdetails.MyListDetailsViewModel
