+++ app/src/main/java/com/elhady/movies/presentation/ui/tvshows/Extensions.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.recyclerview.widget.GridLayoutManager
-import com.elhady.movies.core.common.bases.BaseFooterAdapter
+import com.elhady.movies.core.ui.adapters.BaseFooterAdapter
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
