+++ app/src/main/java/com/elhady/movies/presentation/viewmodel/moviedetails/MovieDetailsViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
 import com.elhady.movies.core.common.bases.BaseViewModel
 import com.elhady.movies.core.common.bases.NavigationRes
-import com.elhady.movies.core.common.bases.StringsRes
+import com.elhady.movies.core.ui.bases.StringsRes
 import com.elhady.movies.core.domain.entities.StatusEntity
 import com.elhady.movies.core.domain.entities.moviedetails.MovieDetailsEntity
 import com.elhady.movies.core.data.ForbiddenThrowable
