+++ app/src/main/java/com/elhady/movies/presentation/viewmodel/mylist/MyListViewModel.kt
@@ -4,7 +4,7 @@ import android.os.Parcel
 import android.os.Parcelable
 import androidx.lifecycle.viewModelScope
 import com.elhady.movies.core.common.bases.BaseViewModel
-import com.elhady.movies.core.common.bases.StringsRes
+import com.elhady.movies.core.ui.bases.StringsRes
 import com.elhady.movies.core.domain.entities.StatusEntity
 import com.elhady.movies.core.domain.usecase.mylist.CreateListUseCase
 import com.elhady.movies.core.domain.usecase.mylist.DeleteListUseCase
