+++ app/src/main/java/com/elhady/movies/presentation/viewmodel/showmore/ShowMoreViewModel.kt
@@ -6,7 +6,7 @@ import androidx.paging.LoadState
 import androidx.paging.PagingData
 import com.elhady.movies.core.common.bases.BaseViewModel
 import com.elhady.movies.core.common.bases.ListType
-import com.elhady.movies.core.common.bases.StringsRes
+import com.elhady.movies.core.ui.bases.StringsRes
 import com.elhady.movies.core.domain.usecase.showmore.GetMorePopularMoviesByTypeUseCase
 import com.elhady.movies.core.domain.usecase.showmore.GetMoreTopRatedByTypeUseCase
 import com.elhady.movies.core.domain.usecase.showmore.GetMoreTrendingByTypeUseCase
